Add support for serverSideData
This data can then been used by the Hypernova server when server-side
rendering a component. The reason why we need this to be separate from
`data` is because Hypernova serializes `data` on the clientside when it
writes a script tag for the hydration process to use.

Therefore we want to ensure that the serverSideData is private and is
never accessible by the browser.

This means we can pass things over to the Hypernova server such as
client headers for making fetch requests (such as the `Authorization`
header on staging environments) without the fear of these being
accessible from Javascript on the client side.